Refactor delay_tween
