+++ lib/src/fading_circle.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_spinkit/src/utils.dart';
+import 'package:flutter_spinkit/src/tweens/delay_tween.dart';
 
 class SpinKitFadingCircle extends StatefulWidget {
   const SpinKitFadingCircle({
